requirements.txt
Flask==3.0.0
requests==2.26.0
Flask-DotEnv==0.1.2
Flask-MySQLdb==2.0.0
